requirements.txt
Python v2.7.15
Pandas-0.3.1
plotly-4.5.0